+++ dongmanee/src/main/java/com/dongmanee/domain/email/service/EmailServiceImpl.java
@@ -10,7 +10,6 @@ import org.springframework.stereotype.Service;
 import com.dongmanee.domain.email.exception.EmailSendingException;
 import com.dongmanee.domain.email.exception.EmailVerifiedException;
 import com.dongmanee.domain.email.utils.EmailRedisUtils;
-import com.dongmanee.domain.member.controller.port.SignUpControllerEmailService;
 import com.dongmanee.domain.member.dao.MemberRepository;
 import com.dongmanee.domain.member.exception.DuplicateEmailException;
 import com.dongmanee.global.utils.AuthCodeProvider;
@@ -25,7 +24,7 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class EmailServiceImpl implements SignUpControllerEmailService {
+public class EmailServiceImpl implements EmailService {
 	private final JavaMailSender emailSender;
 	private final EmailRedisUtils emailRedis;
 	private final MemberRepository memberRepository;
