--- dongmanee/src/main/java/com/dongmanee/domain/member/controller/port/MemberControllerMemberService.java
@@ -1,7 +1,0 @@
-package com.dongmanee.domain.member.controller.port;
-
-import com.dongmanee.domain.member.domain.Member;
-
-public interface MemberControllerMemberService {
-	Member findById(Long id);
-}
