--- dongmanee/src/main/java/com/dongmanee/domain/member/controller/port/SingUpControllerUniversityService.java
@@ -1,9 +1,0 @@
-package com.dongmanee.domain.member.controller.port;
-
-import java.util.Optional;
-
-import com.dongmanee.domain.university.domain.University;
-
-public interface SingUpControllerUniversityService {
-	Optional<University> findById(Long universityId);
-}
