+++ dongmanee/src/main/java/com/dongmanee/domain/member/service/MemberServiceImpl.java
@@ -2,7 +2,6 @@ package com.dongmanee.domain.member.service;
 
 import org.springframework.stereotype.Service;
 
-import com.dongmanee.domain.member.controller.port.MemberControllerMemberService;
 import com.dongmanee.domain.member.dao.MemberRepository;
 import com.dongmanee.domain.member.domain.Member;
 import com.dongmanee.domain.member.exception.MemberNotFoundException;
@@ -11,8 +10,7 @@ import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
-public class MemberServiceImpl
-	implements MemberService, MemberControllerMemberService {
+public class MemberServiceImpl implements MemberService {
 	private final MemberRepository memberRepository;
 
 	@Override
