+++ dongmanee/src/main/java/com/dongmanee/domain/member/service/SignUpService.java
@@ -1,7 +1,7 @@
-package com.dongmanee.domain.member.controller.port;
+package com.dongmanee.domain.member.service;
 
 import com.dongmanee.domain.member.domain.Member;
 
-public interface SignUpControllerSignUpService {
+public interface SignUpService {
 	void signup(String provider, Long externalProviderId, Member member, String emailAuthCode);
 }
